Refactor: Split router/optimizer.py into focused modules
Split the 1326-line optimizer.py into 8 focused modules:
- optimizer/__init__.py (41 lines) - Public API exports
- optimizer/algorithms.py (465 lines) - Core optimization algorithms
- optimizer/chain.py (175 lines) - Chain sorting logic
- optimizer/collision.py (165 lines) - Collision checking
- optimizer/config.py (61 lines) - Configuration dataclasses
- optimizer/geometry.py (200 lines) - Geometry helpers
- optimizer/pcb.py (199 lines) - PCB I/O operations
- optimizer/trace.py (343 lines) - TraceOptimizer class

Each algorithm is now independently testable and the module structure
makes it easy to add new optimization strategies.

Closes #260
